+++ packages/polaris-viz/src/components/TooltipContent/TooltipContent.tsx
@@ -1,6 +1,7 @@
 import {Fragment} from 'react';
-import {DEFAULT_THEME_NAME} from '@shopify/polaris-viz-core';
+import {DEFAULT_THEME_NAME, useChartContext} from '@shopify/polaris-viz-core';
 
+import {hasHiddenComparisonSeries} from '../../utilities/hasHiddenComparisonSeries';
 import type {TooltipData} from '../../types';
 
 import {useGetLongestLabelFromData} from './hooks/useGetLongestLabelFromData';
@@ -30,9 +31,12 @@ export function TooltipContent({
   ignoreColorVisionEvents = false,
 }: TooltipContentProps) {
   const {keyWidth, valueWidth} = useGetLongestLabelFromData(data);
+  const {comparisonSeriesIndexes} = useChartContext();
 
   const leftWidth = keyWidth * FONT_SIZE_OFFSET;
   const rightWidth = valueWidth * FONT_SIZE_OFFSET;
+
+  const isMultiSeries = data[0]?.data.length > 2;
 
   return (
     <TooltipContentContainer
@@ -60,12 +64,29 @@ export function TooltipContent({
                     {key, value, color, isComparison, isHidden},
                     seriesIndex,
                   ) => {
+                    // This check is for when we are rendering multiple series lines
+                    // and we have an active index.
+                    // We only want to render the active series line and it's
+                    // matching comparison series line.
                     if (
-                      // TODO: Check this
-                      data[0].data.length > 2 &&
+                      // If we more than 2 series lines...
+                      isMultiSeries &&
+                      // and the series is not a comparison series...
                       isComparison !== true &&
+                      // and we have an active index...
                       activeColorVisionIndex !== -1 &&
+                      // and the series index is not the active index, hide it.
                       seriesIndex !== activeColorVisionIndex
+                    ) {
+                      return null;
+                    }
+
+                    if (
+                      hasHiddenComparisonSeries({
+                        index: seriesIndex,
+                        activeColorVisionIndex,
+                        comparisonSeriesIndexes,
+                      })
                     ) {
                       return null;
                     }